requirements.txt
asgiref
autopep8
django==2.2.10
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
skpy
mongoengine==0.21.0
django-rest-framework-mongoengine==3.4.1